+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
-## v13 (In Progress)
+## v14 (In Progress)
+- Ancient Cave: Credits roll at end of game (with random music, because why not).
 ### Bugfixes
 - Keysanity: Ness's house lights now behave as normal.
 - Ancient Cave: Further logic errors that caused occasional incorrect enemies to appear in Electro Specter cave areas fixed.